test(e2e_appium): rename the utils package to support
