+++ test/e2e_appium/pages/wallet/keycard_auth_modal.py
@@ -1,5 +1,5 @@
 from locators.wallet.accounts_locators import WalletAccountsLocators
-from utils.exceptions import is_session_fatal
+from support.exceptions import is_session_fatal
 
 from ..base_page import BasePage
