+++ test/e2e_appium/support/chat_state.py
