+++ test/e2e_appium/support/multi_device_helpers.py
