+++ test/e2e_appium/support/platform.py
