+++ test/e2e_appium/tests/messaging/test_emoji_and_media.py
@@ -6,8 +6,8 @@ from config.logging_config import get_logger
 from pages.app import App
 from pages.messaging.chat_page import ChatPage
 from pages.messaging.message_context_menu_page import MessageContextMenuPage
-from utils.generators import generate_account_name
-from utils.timeouts import CROSS_DEVICE_DELIVERY_TIMEOUT_SECONDS
+from support.generators import generate_account_name
+from support.timeouts import CROSS_DEVICE_DELIVERY_TIMEOUT_SECONDS
 
 
 def _unique_message(prefix: str) -> str:
