+++ test/e2e_appium/tests/messaging/test_messaging_1x1_chat.py
@@ -7,7 +7,7 @@ from pages.messaging.chat_page import ChatPage
 from pages.app import App
 from pages.settings.messaging_page import MessagingSettingsPage
 from pages.settings.settings_page import SettingsPage
-from utils.multi_device_helpers import StepMixin
+from support.multi_device_helpers import StepMixin
 
 
 class TestMessaging1x1Chat(StepMixin):
